Only try to load metric explorer charts.
This mitigates the UI lockup when trying to click on a Usage chart from
a report thumbnail (asana issue
https://app.asana.com/0/159049597309611/1153563882786297)

A complete fix will be to actually support all types of charts that can
be shown (Usage, Custom Queries, App Kernel . etc. ). But this at least
ensures that the UI does not lockup.