Fix pie chart datalabel formatting.
The word wrap algorithm for pie chart datalabels would truncate long
lines before word wrapping. The truncation algorithm did not take into
account word boundaries but the wrapping one did. This meant that, even
though the max length is set to the equivalent of two lines of text,
sometimes the result would be on three lines (for a total of four lines
when the values are included). The highchart datalabel layout engine
only leaves enough room for 1 line of text so the charts would be quite
ugly. Furthermore no indication of truncation was given.

For example the text: "Computer and Information Science and Engineering"
would get mangled to:

```
Computer and
Information Science
an
```

The new code runs the word wrap first. If the string spans more than two
lines then ellipses are added to the end of the second line to indicate
that data has been chopped. The padding between datalabels has also been
increased to prevent overlap.

These changes only impact pie charts, but will alter all pie charts in
XDMoD.